Refactor graph visualization and add background image
This commit refactors the graph visualization in the TreeVisualizer class. It updates the positions of the nodes and adds a background image to the canvas. The positions of the nodes have been adjusted to fit the new background image. The SURIGAO_DEL_NORTE_DISTANCE, SURIGAO_DEL_NORTE_COST, and SURIGAO_DEL_NORTE dictionaries have been moved inside the class. The canvas size has also been adjusted to accommodate the new background image.